# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/api/article/add": {
            "post": {
                "description": "添加文章",
                "tags": [
                    "文章"
                ],
                "summary": "添加文章",
                "parameters": [
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Article"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "添加成功返回",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/api/article/get/:id": {
            "get": {
                "description": "查询文章",
                "tags": [
                    "文章"
                ],
                "summary": "查询文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "文章信息",
                        "schema": {
                            "$ref": "#/definitions/model.ArticleResponse"
                        }
                    }
                }
            }
        },
        "/api/article/{id}": {
            "put": {
                "description": "更新文章",
                "tags": [
                    "文章"
                ],
                "summary": "更新文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求体",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.Article"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功返回",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/api/article/{uuid}/status": {
            "put": {
                "description": "更新文章状态",
                "tags": [
                    "文章"
                ],
                "summary": "更新文章状态",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "请求体 (status: 0=Draft, 1=Pending, 2=Published)",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功返回",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/api/articles/delete/:id": {
            "delete": {
                "description": "删除文章",
                "tags": [
                    "文章"
                ],
                "summary": "删除文章",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "文章ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "文章信息",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/comment/add": {
            "post": {
                "description": "添加评论",
                "consumes": [
                    "multipart/form-data"
                ],
                "tags": [
                    "评论"
                ],
                "summary": "添加评论",
                "parameters": [
                    {
                        "type": "string",
                        "description": "评论内容",
                        "name": "content",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "所属帖子ID",
                        "name": "postId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "评论用户ID",
                        "name": "userId",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "description": "父评论ID",
                        "name": "parentId",
                        "in": "formData"
                    },
                    {
                        "type": "integer",
                        "description": "资源类型 image/video",
                        "name": "type",
                        "in": "formData",
                        "required": true
                    },
                    {
                        "type": "file",
                        "description": "上传文件",
                        "name": "files",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"操作成功\"}",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/api/comment/list": {
            "get": {
                "description": "获取评论列表",
                "tags": [
                    "评论"
                ],
                "summary": "获取评论列表",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "帖子ID",
                        "name": "postId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "{\"code\":200,\"data\":{},\"msg\":\"操作成功\"}",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/api/user/add": {
            "post": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "添加用户",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/get/:id": {
            "get": {
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取用户信息",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "用户ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        },
        "/api/user/list": {
            "get": {
                "description": "返回所有用户信息",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "获取用户列表",
                "responses": {
                    "200": {
                        "description": "用户列表",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/model.UserResponse"
                            }
                        }
                    }
                }
            }
        },
        "/api/user/login": {
            "post": {
                "description": "登录",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "登录",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/api/user/update/password": {
            "post": {
                "description": "更新用户密码",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "更新用户密码",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        },
        "/api/user/update/user": {
            "post": {
                "description": "更新用户数据",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "用户"
                ],
                "summary": "更新用户数据",
                "parameters": [
                    {
                        "description": "用户信息",
                        "name": "user",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/model.User"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "成功",
                        "schema": {
                            "$ref": "#/definitions/middleware.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "middleware.Response": {
            "type": "object",
            "properties": {
                "code": {
                    "description": "自定义状态码",
                    "type": "integer"
                },
                "data": {
                    "description": "返回数据"
                },
                "message": {
                    "description": "提示信息",
                    "type": "string"
                },
                "time": {
                    "type": "string"
                }
            }
        },
        "model.Article": {
            "type": "object",
            "properties": {
                "content": {
                    "description": "内容",
                    "type": "string"
                },
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "id": {
                    "description": "主键",
                    "type": "integer"
                },
                "status": {
                    "description": "状态",
                    "enum": [
                        0,
                        1,
                        2
                    ],
                    "allOf": [
                        {
                            "$ref": "#/definitions/model.ArticleStatus"
                        }
                    ]
                },
                "status_name": {
                    "description": "状态名称",
                    "type": "string"
                },
                "title": {
                    "description": "标题",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "user_id": {
                    "description": "外键关联用户ID",
                    "type": "integer"
                }
            }
        },
        "model.ArticleResponse": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "status": {
                    "type": "integer"
                },
                "status_name": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "model.ArticleStatus": {
            "type": "integer",
            "enum": [
                0,
                1,
                2
            ],
            "x-enum-comments": {
                "Draft": "草稿",
                "Pending": "待审核",
                "Published": "已发布"
            },
            "x-enum-varnames": [
                "Draft",
                "Pending",
                "Published"
            ]
        },
        "model.User": {
            "type": "object",
            "required": [
                "password",
                "username"
            ],
            "properties": {
                "created_at": {
                    "description": "创建时间",
                    "type": "string"
                },
                "email": {
                    "description": "邮箱，唯一",
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "password": {
                    "description": "密码",
                    "type": "string"
                },
                "phone": {
                    "description": "手机号",
                    "type": "string"
                },
                "role": {
                    "description": "角色",
                    "type": "string"
                },
                "status": {
                    "description": "状态 active/disabled",
                    "type": "string"
                },
                "updated_at": {
                    "description": "更新时间",
                    "type": "string"
                },
                "username": {
                    "description": "用户名，唯一",
                    "type": "string"
                },
                "uuid": {
                    "description": "用户唯一标识",
                    "type": "string"
                }
            }
        },
        "model.UserResponse": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "deleted_at": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "phone": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                },
                "username": {
                    "type": "string"
                }
            }
        }
    }
}